moving an implementation of a fork of ed25519
